fixed loops, i think

+++ c_backend/functions/bFunctions64/bFunctions64.h
@@ -50,7 +50,7 @@ void mov64_mr_r(uint8_t reg_base, uint8_t reg2, uint8_t reg3);
 
 // Arithmetic operations
 void sum64_r_r(uint8_t reg1, uint8_t reg2);
-void sum64_r_i(uint8_t reg1, uint64_t imm64);
+void sum64_r_i(uint8_t reg1, uint32_t imm64);
 void sum64_r_m(uint8_t reg1, uint8_t reg2);
 void sum64_r_mi(uint8_t reg1, uint8_t reg2, uint64_t offset);
 void sum64_r_mr(uint8_t reg1, uint8_t reg2, uint8_t reg3);
